@@ -21,10 +21,9 @@
 package com.spotify.ffwd.input;
 
 import com.spotify.ffwd.Initializable;
-
 import com.spotify.ffwd.model.v2.Batch;
 import com.spotify.ffwd.model.v2.Metric;
-import eu.toolchain.async.AsyncFuture;
+import java.util.concurrent.CompletableFuture;
 
 public interface InputManager extends Initializable {
 
@@ -35,7 +34,7 @@ public interface InputManager extends Initializable {
 
   void receiveBatch(Batch batch);
 
-  AsyncFuture<Void> start();
+  CompletableFuture<Void> start();
 
-  AsyncFuture<Void> stop();
+  CompletableFuture<Void> stop();
 }
