@@ -20,26 +20,17 @@
 
 package com.spotify.ffwd.debug;
 
-import com.google.inject.Inject;
 import com.spotify.ffwd.model.v2.Batch;
 import com.spotify.ffwd.model.v2.Metric;
 import com.spotify.ffwd.output.BatchablePluginSink;
-import eu.toolchain.async.AsyncFramework;
-import eu.toolchain.async.AsyncFuture;
 import java.util.Collection;
+import java.util.concurrent.CompletableFuture;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 public class DebugPluginSink implements BatchablePluginSink {
 
   private static final Logger log = LoggerFactory.getLogger(DebugPluginSink.class);
-
-  @Inject
-  private AsyncFramework async;
-
-  @Override
-  public void init() {
-  }
 
   @Override
   public void sendMetric(Metric metric) {
@@ -52,39 +43,24 @@ public class DebugPluginSink implements BatchablePluginSink {
   }
 
   @Override
-  public AsyncFuture<Void> sendMetrics(Collection<Metric> metrics) {
+  public CompletableFuture<Void> sendMetrics(Collection<Metric> metrics) {
     int i = 0;
 
     for (final Metric m : metrics) {
       log.info("M#{}: {}", i++, m);
     }
 
-    return async.resolved(null);
+    return CompletableFuture.completedFuture(null);
   }
 
   @Override
-  public AsyncFuture<Void> sendBatches(final Collection<Batch> batches) {
+  public CompletableFuture<Void> sendBatches(final Collection<Batch> batches) {
     int i = 0;
 
     for (final Batch b : batches) {
       log.info("B#{}: {}", i++, b);
     }
 
-    return async.resolved(null);
-  }
-
-  @Override
-  public AsyncFuture<Void> start() {
-    return async.resolved();
-  }
-
-  @Override
-  public AsyncFuture<Void> stop() {
-    return async.resolved();
-  }
-
-  @Override
-  public boolean isReady() {
-    return true;
+    return CompletableFuture.completedFuture(null);
   }
 }
