@@ -20,16 +20,12 @@
 
 package com.spotify.ffwd.protocol;
 
-import eu.toolchain.async.AsyncFramework;
-import eu.toolchain.async.AsyncFuture;
-import eu.toolchain.async.ResolvableFuture;
 import io.netty.channel.Channel;
 import io.netty.channel.ChannelFuture;
 import io.netty.channel.ChannelFutureListener;
-import io.netty.util.Timeout;
 import io.netty.util.Timer;
-import io.netty.util.TimerTask;
 import java.util.Collection;
+import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
@@ -42,25 +38,22 @@ public class RetryingProtocolConnection implements ProtocolConnection {
   private final AtomicReference<Channel> channel = new AtomicReference<>();
   private final Object lock = new Object();
 
-  private final AsyncFramework async;
   private final Timer timer;
   private final Logger log;
   private final RetryPolicy policy;
   private final ProtocolChannelSetup action;
 
-  private final ResolvableFuture<ProtocolConnection> initialFuture;
+  private final CompletableFuture<ProtocolConnection> initialFuture;
 
-  public RetryingProtocolConnection(
-      AsyncFramework async, Timer timer, Logger log, RetryPolicy policy,
-      ProtocolChannelSetup action
+  RetryingProtocolConnection(
+      Timer timer, Logger log, RetryPolicy policy, ProtocolChannelSetup action
   ) {
-    this.async = async;
     this.timer = timer;
     this.log = log;
     this.policy = policy;
     this.action = action;
 
-    this.initialFuture = async.<ProtocolConnection>future();
+    this.initialFuture = new CompletableFuture<>();
 
     trySetup(0);
   }
@@ -70,32 +63,26 @@ public class RetryingProtocolConnection implements ProtocolConnection {
 
     final ChannelFuture connect = action.setup();
 
-    connect.addListener(new ChannelFutureListener() {
-      @Override
-      public void operationComplete(ChannelFuture future) throws Exception {
-        if (future.isSuccess()) {
-          log.info("Successful {}", action);
-          setChannel(future.channel());
+    connect.addListener((ChannelFutureListener) future -> {
+      if (future.isSuccess()) {
+        log.info("Successful {}", action);
+        setChannel(future.channel());
+        return;
+      }
+
+      final long delay = policy.delay(attempt);
+
+      log.warn("Failed {} (attempt: {}), retrying in {}s: {}", action, attempt + 1,
+          TimeUnit.SECONDS.convert(delay, TimeUnit.MILLISECONDS),
+          future.cause().getMessage());
+
+      timer.newTimeout(timeout -> {
+        if (stopped.get()) {
           return;
         }
 
-        final long delay = policy.delay(attempt);
-
-        log.warn("Failed {} (attempt: {}), retrying in {}s: {}", action, attempt + 1,
-            TimeUnit.SECONDS.convert(delay, TimeUnit.MILLISECONDS),
-            future.cause().getMessage());
-
-        timer.newTimeout(new TimerTask() {
-          @Override
-          public void run(Timeout timeout) throws Exception {
-            if (stopped.get()) {
-              return;
-            }
-
-            trySetup(attempt + 1);
-          }
-        }, delay, TimeUnit.MILLISECONDS);
-      }
+        trySetup(attempt + 1);
+      }, delay, TimeUnit.MILLISECONDS);
     });
   }
 
@@ -110,24 +97,21 @@ public class RetryingProtocolConnection implements ProtocolConnection {
       }
 
       if (!initialFuture.isDone()) {
-        initialFuture.resolve(this);
+        initialFuture.complete(this);
       }
 
       channel.set(c);
     }
 
-    c.closeFuture().addListener(new ChannelFutureListener() {
-      @Override
-      public void operationComplete(ChannelFuture future) throws Exception {
-        log.info("Lost {}, retrying", action);
-        channel.set(null);
-        trySetup(0);
-      }
+    c.closeFuture().addListener((ChannelFutureListener) future -> {
+      log.info("Lost {}, retrying", action);
+      channel.set(null);
+      trySetup(0);
     });
   }
 
   @Override
-  public AsyncFuture<Void> stop() {
+  public CompletableFuture<Void> stop() {
     final Channel c;
 
     synchronized (lock) {
@@ -136,20 +120,17 @@ public class RetryingProtocolConnection implements ProtocolConnection {
       c = channel.getAndSet(null);
 
       if (c == null) {
-        return async.resolved(null);
+        return CompletableFuture.completedFuture(null);
       }
     }
 
-    final ResolvableFuture<Void> future = async.future();
+    CompletableFuture<Void> future = new CompletableFuture<>();
 
-    c.close().addListener(new ChannelFutureListener() {
-      @Override
-      public void operationComplete(ChannelFuture f) throws Exception {
-        try {
-          future.resolve(f.get());
-        } catch (ExecutionException e) {
-          future.fail(e);
-        }
+    c.close().addListener((ChannelFutureListener) f -> {
+      try {
+        future.complete(f.get());
+      } catch (ExecutionException e) {
+        future.completeExceptionally(e);
       }
     });
 
@@ -164,34 +145,28 @@ public class RetryingProtocolConnection implements ProtocolConnection {
       return;
     }
 
-    c.writeAndFlush(message).addListener(new ChannelFutureListener() {
-      @Override
-      public void operationComplete(ChannelFuture future) throws Exception {
-        if (!future.isSuccess()) {
-          log.error("failed to send metric", future.cause());
-        }
+    c.writeAndFlush(message).addListener((ChannelFutureListener) future -> {
+      if (!future.isSuccess()) {
+        log.error("failed to send metric", future.cause());
       }
     });
   }
 
   @Override
-  public AsyncFuture<Void> sendAll(Collection<? extends Object> batch) {
+  public CompletableFuture<Void> sendAll(Collection<? extends Object> batch) {
     final Channel c = channel.get();
 
     if (c == null) {
-      return async.failed(new IllegalStateException("not connected"));
+      return CompletableFuture.failedFuture(new IllegalStateException("not connected"));
     }
 
-    final ResolvableFuture<Void> future = async.future();
+    CompletableFuture<Void> future = new CompletableFuture<>();
 
-    c.writeAndFlush(batch).addListener(new ChannelFutureListener() {
-      @Override
-      public void operationComplete(ChannelFuture f) throws Exception {
-        try {
-          future.resolve(f.get());
-        } catch (ExecutionException e) {
-          future.fail(e);
-        }
+    c.writeAndFlush(batch).addListener((ChannelFutureListener) f -> {
+      try {
+        future.complete(f.get());
+      } catch (ExecutionException e) {
+        future.completeExceptionally(e);
       }
     });
 
@@ -212,7 +187,7 @@ public class RetryingProtocolConnection implements ProtocolConnection {
   /**
    * Return a future that will be resolved when an initial action has been successful.
    */
-  public AsyncFuture<ProtocolConnection> getInitialFuture() {
+  public CompletableFuture<ProtocolConnection> getInitialFuture() {
     return initialFuture;
   }
 }
