@@ -22,15 +22,11 @@ package com.spotify.ffwd.http;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.netflix.loadbalancer.Server;
-import eu.toolchain.async.AsyncFramework;
 import javax.inject.Inject;
 import javax.inject.Named;
 import okhttp3.OkHttpClient;
 
 public class HttpClientFactory {
-
-  @Inject
-  private AsyncFramework async;
 
   @Inject
   @Named("application/json")
@@ -41,6 +37,6 @@ public class HttpClientFactory {
 
   public HttpClient newClient(final Server server) {
     final String baseUrl = "http://" + server.getHost() + ":" + server.getPort();
-    return new HttpClient(async, mapper, httpClient, baseUrl);
+    return new HttpClient(mapper, httpClient, baseUrl);
   }
 }
